+++ Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/Utility/SavableImage.swift
@@ -19,21 +19,6 @@
 
 import Photos
 
-protocol ImageManagerProtocol {
-    func cancelImageRequest(_ requestID: PHImageRequestID)
-
-    @discardableResult
-    func requestImage(for asset: PHAsset, targetSize: CGSize, contentMode: PHImageContentMode, options: PHImageRequestOptions?, resultHandler: @escaping (UIImage?, [AnyHashable : Any]?) -> Void) -> PHImageRequestID
-
-    @discardableResult
-    func requestImageData(for asset: PHAsset, options: PHImageRequestOptions?, resultHandler: @escaping (Data?, String?, UIImage.Orientation, [AnyHashable : Any]?) -> Void) -> PHImageRequestID
-
-    @discardableResult
-    func requestExportSession(forVideo asset: PHAsset, options: PHVideoRequestOptions?, exportPreset: String, resultHandler: @escaping (AVAssetExportSession?, [AnyHashable : Any]?) -> Void) -> PHImageRequestID
-}
-
-extension PHImageManager: ImageManagerProtocol {}
-
 protocol PhotoLibraryProtocol {
     func performChanges(_ changeBlock: @escaping () -> Swift.Void, completionHandler: ((Bool, Error?) -> Swift.Void)?)
